fix: Fix validation for FieldGroup and collection field selection
Fixes: #844
Fixes: #860
