+++ lib/itests/core/src/test/resources/log4j2.xml
@@ -9,7 +9,7 @@
     </File>
   </Appenders>
   <Loggers>
-    <Logger name="io.atlasmap" level="debug" additivity="false">
+    <Logger name="io.atlasmap" level="info" additivity="false">
       <AppenderRef ref="Console" />
       <AppenderRef ref="File" />
     </Logger>
