+++ lib/modules/json/module/src/main/java/io/atlasmap/json/module/JsonValidationService.java
@@ -19,7 +19,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import io.atlasmap.core.BaseModuleValidationService;
+import io.atlasmap.core.validate.BaseModuleValidationService;
 import io.atlasmap.json.v2.JsonField;
 import io.atlasmap.spi.AtlasConversionService;
 import io.atlasmap.spi.AtlasFieldActionService;
